Moves create Thought request to the dedicated directory
