+++ app/Http/Controllers/Api/Thoughts/CreateThoughtController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Api\Thoughts;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreThoughtRequest;
+use App\Http\Requests\Thoughts\StoreThoughtRequest;
 use App\Models\Thought;
 use Illuminate\Http\JsonResponse;
