+++ app/Http/Requests/Thoughts/StoreThoughtRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Thoughts;
 
 use Illuminate\Foundation\Http\FormRequest;
